pyproject.toml
[tool.poetry]
name = "e-commerce"
version = "0.1.0"
description = ""
authors = ["dedoxd2 <dedo.xd33@yahoo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.1"
python-decouple = "^3.8"
pytest = "^7.0.0"
pytest-base-url = "^2.0.0"
pytest-django = "^4.8.0"
pytest-factoryboy = "^2.6.0"
pytest-selenium = "^4.0.2"
black = "^24.1.1"
django-mptt = "^0.16.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
